chore: rename controller checkup
